Moves SortOrder type to interfaces file
